Push detailed documentation into sectiondefs
In preparation to support user-defined sections, add a new
to_details_asciidoc function to each sectiondef class that returns the
AsciiDoc with detailed documentation for that section. This way, it
makes it easy for us to tighten up the documentation so only immediate
children memberdefs of a sectiondef are documented (rather than
hoovering up memberdefs anywhere in a group which might duplicate
documentation from a user-defined section).