Pass an explicit depth when rendering nodes
Rather than having a mix of depths and hard-coded levels for certain
titles, rely entirely on a `depth` keyword argument passed to
`to_asciidoc` which dictates the current level of any node.

It is a parent node's responsibility to increase the depth when
rendering a child that introduces a new level (e.g. a detailed
description inside a module).

This way, we can render the Doxygen index starting at depth 2 and any
arbitrary pages (e.g. the Examples page) at depth 1.

This also fixes the level of `<title>`s within `<sect1>`s (e.g. the
"Interrupt Numbers" section of hardware_irq).