Expand usage of argparse to handle CLI options
Rather than using optional string arguments for both input and output files,
use more of Python's argparse library to handle a single, mandatory,
positional file argument (which now returns a handle to read the file).

This is a breaking change as the "-f" argument is now a positional one
so any downstream usage will need to be updated.

Note we don't use `argparse.FileType("w")` for the output file as that
would cause the given file to be created or truncated as soon as the
program runs when we only want it to produce output if successful.