Refactor attribute handling
Rather than passing around an AsciiDoc string of attributes, pass around
a dictionary that is only converted to AsciiDoc at the very end when
passed to the `title` helper.

Note I considered switching to using the role shorthand for multiple
roles (e.g. `.h6.contextspecific` rather than `role=h6 contextspecific`)
but there is downstream code that relies on matching the existing
format. Similarly, the order of the `tag` and `type` attributes is
currently relied upon by client code so these have to be special cased
for now.

This also removes the dedicated Title Node class in favour of handling
title rendering in the parent element (in this case, the Sect Node).